# Conflicts:
#	components/ax-cgroup/src/lib.rs
#	components/ax-cgroup/src/membership.rs
#	components/ax-cgroup/src/node.rs
#	os/StarryOS/kernel/src/cgroup/mod.rs
#	os/StarryOS/kernel/src/syscall/fs/ctl.rs
#	os/StarryOS/kernel/src/syscall/signal.rs
#	os/StarryOS/kernel/src/syscall/sys.rs
#	os/StarryOS/kernel/src/syscall/task/clone.rs
#	os/StarryOS/kernel/src/syscall/task/clone3.rs
#	os/StarryOS/kernel/src/syscall/task/ctl.rs
#	os/StarryOS/kernel/src/syscall/task/namespace.rs
#	os/StarryOS/kernel/src/task/ops.rs

+++ os/StarryOS/kernel/src/syscall/mm/mmap.rs
@@ -680,7 +680,7 @@ pub fn sys_mprotect(
     Ok(0)
 }
 
-const MREMAP_VALID_FLAGS: u32 = MREMAP_MAYMOVE | MREMAP_FIXED | MREMAP_DONTUNMAP;
+const MREMAP_VALID_FLAGS: usize = (MREMAP_MAYMOVE | MREMAP_FIXED | MREMAP_DONTUNMAP) as usize;
 
 fn find_free(
     aspace: &crate::mm::AddrSpace,
@@ -782,7 +782,7 @@ pub fn sys_mremap(
     addr: usize,
     old_size: usize,
     new_size: usize,
-    flags: u32,
+    flags: usize,
     new_addr: usize,
 ) -> StarryResult<isize> {
     debug!(
@@ -798,9 +798,9 @@ pub fn sys_mremap(
     }
 
     let addr = VirtAddr::from(addr);
-    let may_move = flags & MREMAP_MAYMOVE != 0;
-    let fixed = flags & MREMAP_FIXED != 0;
-    let dontunmap = flags & MREMAP_DONTUNMAP != 0;
+    let may_move = flags & MREMAP_MAYMOVE as usize != 0;
+    let fixed = flags & MREMAP_FIXED as usize != 0;
+    let dontunmap = flags & MREMAP_DONTUNMAP as usize != 0;
 
     if (fixed || dontunmap) && !may_move {
         return Err(StarryError::InvalidInput);
